[新增功能](dev_2.3.1): 需求：1643 新增365卡产品售出激活统计表（运营）
1、增加接口：根据条件查询365卡产品售出激活数据统计表
2、增加接口：根据条件导出365卡产品售出激活数据统计表
3、增加接口：根据条件查询365卡产品售出明细统计表
4、增加接口：根据条件导出365卡产品售出明细统计表
5、增加接口：根据条件查询365卡产品激活明细统计表
6、增加接口：根据条件导出365卡产品激活明细统计表

+++ yunya-remote/src/main/java/com/yunya/feign/report/domain/query/BaseCouponQueryForm.java
@@ -32,7 +32,7 @@ public class BaseCouponQueryForm extends PageQuery implements Serializable {
 
     /** 产品类型id列表 */
     @ApiModelProperty("产品类型id列表")
-    private Collection<Integer> productTypeIds;
+    private Collection<Byte> productTypeIds;
 
     /** 产品名称 */
     @ApiModelProperty("产品名称")
